+++ src/OctoshiftCLI.Tests/Octoshift/Handlers/WaitForMigrationCommandHandlerTests.cs
@@ -18,6 +18,7 @@ public class WaitForMigrationCommandHandlerTests
     private const string REPO_MIGRATION_ID = "RM_MIGRATION_ID";
     private const string ORG_MIGRATION_ID = "OM_MIGRATION_ID";
     private const string TARGET_REPO = "TARGET_REPO";
+    private const string MIGRATION_URL = "URL";
     private const int WAIT_INTERVAL = 0;
 
     public WaitForMigrationCommandHandlerTests()
@@ -33,9 +34,10 @@ public class WaitForMigrationCommandHandlerTests
     {
         // Arrange
         _mockGithubApi.SetupSequence(x => x.GetMigration(REPO_MIGRATION_ID).Result)
-            .Returns((State: RepositoryMigrationStatus.InProgress, RepositoryName: TARGET_REPO, FailureReason: null))
-            .Returns((State: RepositoryMigrationStatus.InProgress, RepositoryName: TARGET_REPO, FailureReason: null))
-            .Returns((State: RepositoryMigrationStatus.Succeeded, RepositoryName: TARGET_REPO, FailureReason: null));
+            .Returns((State: RepositoryMigrationStatus.InProgress, RepositoryName: TARGET_REPO, FailureReason: null, MigrationLogUrl: MIGRATION_URL))
+            .Returns((State: RepositoryMigrationStatus.InProgress, RepositoryName: TARGET_REPO, FailureReason: null, MigrationLogUrl: MIGRATION_URL))
+            .Returns((State: RepositoryMigrationStatus.Succeeded, RepositoryName: TARGET_REPO, FailureReason: null, MigrationLogUrl: MIGRATION_URL));
+        _mockGithubApi.Setup(x => x.GetMigrationLogUrl(It.IsAny<string>(), It.IsAny<string>())).ReturnsAsync(MIGRATION_URL);
 
         var actualLogOutput = new List<string>();
         _mockOctoLogger.Setup(m => m.LogInformation(It.IsAny<string>())).Callback<string>(s => actualLogOutput.Add(s));
@@ -48,8 +50,9 @@ public class WaitForMigrationCommandHandlerTests
                 $"Waiting {WAIT_INTERVAL} seconds...",
                 $"Migration {REPO_MIGRATION_ID} for {TARGET_REPO} is {RepositoryMigrationStatus.InProgress}",
                 $"Waiting {WAIT_INTERVAL} seconds...",
-                $"Migration {REPO_MIGRATION_ID} succeeded for {TARGET_REPO}"
-            };
+                $"Migration {REPO_MIGRATION_ID} succeeded for {TARGET_REPO}",
+                $"Migration log available at {MIGRATION_URL} or by running `gh {CliContext.RootCommand} download-logs`"
+    };
 
         // Act
         var args = new WaitForMigrationCommandArgs
@@ -59,7 +62,7 @@ public class WaitForMigrationCommandHandlerTests
         await _handler.Handle(args);
 
         // Assert
-        _mockOctoLogger.Verify(m => m.LogInformation(It.IsAny<string>()), Times.Exactly(5));
+        _mockOctoLogger.Verify(m => m.LogInformation(It.IsAny<string>()), Times.Exactly(6));
         _mockOctoLogger.Verify(m => m.LogSuccess(It.IsAny<string>()), Times.Once);
 
         _mockGithubApi.Verify(m => m.GetMigration(REPO_MIGRATION_ID), Times.Exactly(3));
@@ -76,9 +79,10 @@ public class WaitForMigrationCommandHandlerTests
         const string failureReason = "FAILURE_REASON";
 
         _mockGithubApi.SetupSequence(x => x.GetMigration(REPO_MIGRATION_ID).Result)
-            .Returns((State: RepositoryMigrationStatus.InProgress, RepositoryName: TARGET_REPO, FailureReason: null))
-            .Returns((State: RepositoryMigrationStatus.InProgress, RepositoryName: TARGET_REPO, FailureReason: null))
-            .Returns((State: RepositoryMigrationStatus.Failed, RepositoryName: TARGET_REPO, FailureReason: failureReason));
+            .Returns((State: RepositoryMigrationStatus.InProgress, RepositoryName: TARGET_REPO, FailureReason: null, MigrationLogUrl: MIGRATION_URL))
+            .Returns((State: RepositoryMigrationStatus.InProgress, RepositoryName: TARGET_REPO, FailureReason: null, MigrationLogUrl: MIGRATION_URL))
+            .Returns((State: RepositoryMigrationStatus.Failed, RepositoryName: TARGET_REPO, FailureReason: failureReason, MigrationLogUrl: MIGRATION_URL));
+        _mockGithubApi.Setup(x => x.GetMigrationLogUrl(It.IsAny<string>(), It.IsAny<string>())).ReturnsAsync(MIGRATION_URL);
 
         var actualLogOutput = new List<string>();
         _mockOctoLogger.Setup(m => m.LogInformation(It.IsAny<string>())).Callback<string>(s => actualLogOutput.Add(s));
@@ -91,7 +95,8 @@ public class WaitForMigrationCommandHandlerTests
                 $"Waiting {WAIT_INTERVAL} seconds...",
                 $"Migration {REPO_MIGRATION_ID} for {TARGET_REPO} is {RepositoryMigrationStatus.InProgress}",
                 $"Waiting {WAIT_INTERVAL} seconds...",
-                $"Migration {REPO_MIGRATION_ID} failed for {TARGET_REPO}"
+                $"Migration {REPO_MIGRATION_ID} failed for {TARGET_REPO}",
+                $"Migration log available at {MIGRATION_URL} or by running `gh {CliContext.RootCommand} download-logs`"
             };
 
         // Act
@@ -106,7 +111,7 @@ public class WaitForMigrationCommandHandlerTests
             .WithMessage(failureReason);
 
         // Assert
-        _mockOctoLogger.Verify(m => m.LogInformation(It.IsAny<string>()), Times.Exactly(5));
+        _mockOctoLogger.Verify(m => m.LogInformation(It.IsAny<string>()), Times.Exactly(6));
         _mockOctoLogger.Verify(m => m.LogError(It.IsAny<string>()), Times.Once);
 
         _mockGithubApi.Verify(m => m.GetMigration(REPO_MIGRATION_ID), Times.Exactly(3));
@@ -123,9 +128,10 @@ public class WaitForMigrationCommandHandlerTests
         const string failureReason = "FAILURE_REASON";
 
         _mockGithubApi.SetupSequence(x => x.GetMigration(REPO_MIGRATION_ID).Result)
-            .Returns((State: RepositoryMigrationStatus.PendingValidation, RepositoryName: TARGET_REPO, FailureReason: null))
-            .Returns((State: RepositoryMigrationStatus.PendingValidation, RepositoryName: TARGET_REPO, FailureReason: null))
-            .Returns((State: RepositoryMigrationStatus.FailedValidation, RepositoryName: TARGET_REPO, FailureReason: failureReason));
+            .Returns((State: RepositoryMigrationStatus.PendingValidation, RepositoryName: TARGET_REPO, FailureReason: null, MigrationLogUrl: MIGRATION_URL))
+            .Returns((State: RepositoryMigrationStatus.PendingValidation, RepositoryName: TARGET_REPO, FailureReason: null, MigrationLogUrl: MIGRATION_URL))
+            .Returns((State: RepositoryMigrationStatus.FailedValidation, RepositoryName: TARGET_REPO, FailureReason: failureReason, MigrationLogUrl: MIGRATION_URL));
+        _mockGithubApi.Setup(x => x.GetMigrationLogUrl(It.IsAny<string>(), It.IsAny<string>())).ReturnsAsync(MIGRATION_URL);
 
         var actualLogOutput = new List<string>();
         _mockOctoLogger.Setup(m => m.LogInformation(It.IsAny<string>())).Callback<string>(s => actualLogOutput.Add(s));
@@ -138,7 +144,8 @@ public class WaitForMigrationCommandHandlerTests
                 $"Waiting {WAIT_INTERVAL} seconds...",
                 $"Migration {REPO_MIGRATION_ID} for {TARGET_REPO} is {RepositoryMigrationStatus.PendingValidation}",
                 $"Waiting {WAIT_INTERVAL} seconds...",
-                $"Migration {REPO_MIGRATION_ID} failed for {TARGET_REPO}"
+                $"Migration {REPO_MIGRATION_ID} failed for {TARGET_REPO}",
+                $"Migration log available at {MIGRATION_URL} or by running `gh {CliContext.RootCommand} download-logs`"
             };
 
         // Act
@@ -154,7 +161,7 @@ public class WaitForMigrationCommandHandlerTests
 
         // Assert
 
-        _mockOctoLogger.Verify(m => m.LogInformation(It.IsAny<string>()), Times.Exactly(5));
+        _mockOctoLogger.Verify(m => m.LogInformation(It.IsAny<string>()), Times.Exactly(6));
         _mockOctoLogger.Verify(m => m.LogError(It.IsAny<string>()), Times.Once);
 
         _mockGithubApi.Verify(m => m.GetMigration(REPO_MIGRATION_ID), Times.Exactly(3));