+++ src/bbs2gh/Commands/GenerateScriptCommand.cs
@@ -23,13 +23,16 @@ public class GenerateScriptCommand : CommandBase<GenerateScriptCommandArgs, Gene
         AddOption(SshUser);
         AddOption(SshPrivateKey);
         AddOption(SshPort);
+        AddOption(ArchiveDownloadHost);
         AddOption(SmbUser);
         AddOption(SmbDomain);
         AddOption(Output);
         AddOption(Kerberos);
         AddOption(Verbose);
         AddOption(AwsBucketName);
+        AddOption(AwsRegion);
         AddOption(KeepArchive);
+        AddOption(NoSslVerify);
     }
 
     public Option<string> BbsServerUrl { get; } = new(
@@ -55,6 +58,11 @@ public class GenerateScriptCommand : CommandBase<GenerateScriptCommandArgs, Gene
         name: "--bbs-shared-home",
         description: "Bitbucket server's shared home directory. Defaults to \"/var/atlassian/application-data/bitbucket/shared\" if downloading the archive from a server using SSH " +
                      "and \"c$\\atlassian\\applicationdata\\bitbucket\\shared\" if downloading using SMB.");
+
+    public Option<string> ArchiveDownloadHost { get; } = new(
+        name: "--archive-download-host",
+        description: "The host to use to connect to the Bitbucket Server/Data Center instance via SSH or SMB. Defaults to the host from the Bitbucket Server URL (--bbs-server-url).")
+    { IsHidden = true };
 
     public Option<string> SshUser { get; } = new(
         name: "--ssh-user",
@@ -95,11 +103,22 @@ public class GenerateScriptCommand : CommandBase<GenerateScriptCommandArgs, Gene
         name: "--aws-bucket-name",
         description: "If using AWS, the name of the S3 bucket to upload the BBS archive to.");
 
+    public Option<string> AwsRegion { get; } = new(
+        name: "--aws-region",
+        description: "If using AWS, the AWS region. If not provided, it will be read from AWS_REGION environment variable. " +
+                     "Defaults to us-east-1 if neither the argument nor the environment variable is set. " +
+                     "In a future release, you will be required to set an AWS region if using AWS S3 as your blob storage provider.");
+
     public Option<bool> Verbose { get; } = new("--verbose");
 
     public Option<bool> KeepArchive { get; } = new(
         name: "--keep-archive",
         description: "Keeps the downloaded export archive after successfully uploading it. By default, it will be automatically deleted.");
+
+    public Option<bool> NoSslVerify { get; } = new(
+        name: "--no-ssl-verify",
+        description: "Disables SSL verification when communicating with your Bitbucket Server/Data Center instance. All other migration steps will continue to verify SSL. " +
+                     "If your Bitbucket instance has a self-signed SSL certificate then setting this flag will allow the migration archive to be exported.");
 
     public override GenerateScriptCommandHandler BuildHandler(GenerateScriptCommandArgs args, IServiceProvider sp)
     {
@@ -119,7 +138,9 @@ public class GenerateScriptCommand : CommandBase<GenerateScriptCommandArgs, Gene
         var environmentVariableProvider = sp.GetRequiredService<EnvironmentVariableProvider>();
 
         var bbsApiFactory = sp.GetRequiredService<BbsApiFactory>();
-        var bbsApi = args.Kerberos ? bbsApiFactory.CreateKerberos(args.BbsServerUrl) : bbsApiFactory.Create(args.BbsServerUrl, args.BbsUsername, args.BbsPassword);
+        var bbsApi = args.Kerberos
+            ? bbsApiFactory.CreateKerberos(args.BbsServerUrl, args.NoSslVerify)
+            : bbsApiFactory.Create(args.BbsServerUrl, args.BbsUsername, args.BbsPassword, args.NoSslVerify);
 
         return new GenerateScriptCommandHandler(log, versionProvider, fileSystemProvider, bbsApi, environmentVariableProvider);
     }
@@ -133,6 +154,7 @@ public class GenerateScriptCommandArgs
     public string BbsPassword { get; set; }
     public string BbsProjectKey { get; set; }
     public string BbsSharedHome { get; set; }
+    public string ArchiveDownloadHost { get; set; }
     public string SshUser { get; set; }
     public string SshPrivateKey { get; set; }
     public int SshPort { get; set; }
@@ -142,5 +164,7 @@ public class GenerateScriptCommandArgs
     public bool Kerberos { get; set; }
     public bool Verbose { get; set; }
     public string AwsBucketName { get; set; }
+    public string AwsRegion { get; set; }
     public bool KeepArchive { get; set; }
+    public bool NoSslVerify { get; set; }
 }