+++ src/bbs2gh/Handlers/MigrateRepoCommandHandler.cs
@@ -52,6 +52,9 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         _log.RegisterSecret(args.BbsPassword);
         _log.RegisterSecret(args.GithubPat);
         _log.RegisterSecret(args.SmbPassword);
+        _log.RegisterSecret(args.AwsAccessKey);
+        _log.RegisterSecret(args.AwsSecretKey);
+        _log.RegisterSecret(args.AwsSessionToken);
 
         LogOptions(args);
         ValidateOptions(args);
@@ -231,21 +234,23 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
             return;
         }
 
-        var (migrationState, _, failureReason) = await _githubApi.GetMigration(migrationId);
+        var (migrationState, _, failureReason, migrationLogUrl) = await _githubApi.GetMigration(migrationId);
 
         while (RepositoryMigrationStatus.IsPending(migrationState))
         {
             _log.LogInformation($"Migration in progress (ID: {migrationId}). State: {migrationState}. Waiting 10 seconds...");
             await Task.Delay(CHECK_STATUS_DELAY_IN_MILLISECONDS);
-            (migrationState, _, failureReason) = await _githubApi.GetMigration(migrationId);
+            (migrationState, _, failureReason, migrationLogUrl) = await _githubApi.GetMigration(migrationId);
         }
 
         if (RepositoryMigrationStatus.IsFailed(migrationState))
         {
+            _log.LogInformation($"Migration log available at {migrationLogUrl} or by running `gh {CliContext.RootCommand} download-logs --github-target-org {args.GithubOrg} --target-repo {args.GithubRepo}`");
             throw new OctoshiftCliException($"Migration #{migrationId} failed: {failureReason}");
         }
 
         _log.LogSuccess($"Migration completed (ID: {migrationId})! State: {migrationState}");
+        _log.LogInformation($"Migration log available at {migrationLogUrl} or by running `gh {CliContext.RootCommand} download-logs --github-target-org {args.GithubOrg} --target-repo {args.GithubRepo}`");
     }
 
     private void LogOptions(MigrateRepoCommandArgs args)
@@ -292,20 +297,7 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
             _log.LogInformation($"AZURE STORAGE CONNECTION STRING: ********");
         }
 
-        if (args.AwsBucketName.HasValue())
-        {
-            _log.LogInformation($"AWS BUCKET NAME: {args.AwsBucketName}");
-        }
-
-        if (args.AwsAccessKey.HasValue())
-        {
-            _log.LogInformation($"AWS ACCESS KEY: ********");
-        }
-
-        if (args.AwsSecretKey.HasValue())
-        {
-            _log.LogInformation($"AWS SECRET KEY: ********");
-        }
+        LogAwsOptions(args);
 
         if (args.GithubOrg.HasValue())
         {
@@ -315,6 +307,11 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         if (args.GithubRepo.HasValue())
         {
             _log.LogInformation($"GITHUB REPO: {args.GithubRepo}");
+        }
+
+        if (args.SshUser.HasValue())
+        {
+            _log.LogInformation($"ARCHIVE DOWNLOAD HOST: {args.ArchiveDownloadHost}");
         }
 
         if (args.SshUser.HasValue())
@@ -366,6 +363,39 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         {
             _log.LogInformation("KEEP ARCHIVE: true");
         }
+
+        if (args.NoSslVerify)
+        {
+            _log.LogInformation("NO SSL VERIFY: true");
+        }
+    }
+
+    private void LogAwsOptions(MigrateRepoCommandArgs args)
+    {
+        if (args.AwsBucketName.HasValue())
+        {
+            _log.LogInformation($"AWS BUCKET NAME: {args.AwsBucketName}");
+        }
+
+        if (args.AwsAccessKey.HasValue())
+        {
+            _log.LogInformation("AWS ACCESS KEY: ********");
+        }
+
+        if (args.AwsSecretKey.HasValue())
+        {
+            _log.LogInformation("AWS SECRET KEY: ********");
+        }
+
+        if (args.AwsSessionToken.HasValue())
+        {
+            _log.LogInformation("AWS SESSION TOKEN: ********");
+        }
+
+        if (args.AwsRegion.HasValue())
+        {
+            _log.LogInformation($"AWS REGION: {args.AwsRegion}");
+        }
     }
 
     private void ValidateOptions(MigrateRepoCommandArgs args)
@@ -412,10 +442,7 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
                 }
             }
 
-            if (ShouldDownloadArchive(args))
-            {
-                ValidateDownloadOptions(args);
-            }
+            ValidateDownloadOptions(args);
         }
         else
         {
@@ -424,7 +451,12 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
                 throw new OctoshiftCliException("--bbs-username and --bbs-password can only be provided with --bbs-server-url.");
             }
 
-            if (new[] { args.SshUser, args.SshPrivateKey, args.SmbUser, args.SmbPassword, args.SmbDomain }.Any(obj => obj.HasValue()))
+            if (args.NoSslVerify)
+            {
+                throw new OctoshiftCliException("--no-ssl-verify can only be provided with --bbs-server-url.");
+            }
+
+            if (new[] { args.SshUser, args.SshPrivateKey, args.ArchiveDownloadHost, args.SmbUser, args.SmbPassword, args.SmbDomain }.Any(obj => obj.HasValue()))
             {
                 throw new OctoshiftCliException("SSH or SMB download options can only be provided with --bbs-server-url.");
             }
@@ -462,6 +494,11 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         {
             throw new OctoshiftCliException("Both --smb-user and --smb-password (or SMB_PASSWORD env. variable) must be specified for SMB download.");
         }
+
+        if (args.ArchiveDownloadHost.HasValue() && !shouldUseSsh && !shouldUseSmb)
+        {
+            throw new OctoshiftCliException("--archive-download-host can only be provided if SSH or SMB download options are provided.");
+        }
     }
 
     private void ValidateUploadOptions(MigrateRepoCommandArgs args)
@@ -472,7 +509,7 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         {
             throw new OctoshiftCliException(
                 "Either Azure storage connection (--azure-storage-connection-string or AZURE_STORAGE_CONNECTION_STRING env. variable) or " +
-                "AWS S3 connection (--aws-bucket-name, --aws-access-key (or AWS_ACCESS_KEY env. variable), --aws-secret-key (or AWS_SECRET_Key env.variable)) " +
+                "AWS S3 connection (--aws-bucket-name, --aws-access-key (or AWS_ACCESS_KEY_ID env. variable), --aws-secret-key (or AWS_SECRET_ACCESS_KEY env.variable)) " +
                 "must be provided.");
         }
 
@@ -480,7 +517,7 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         {
             throw new OctoshiftCliException(
                 "Azure storage connection (--azure-storage-connection-string or AZURE_STORAGE_CONNECTION_STRING env. variable) and " +
-                "AWS S3 connection (--aws-bucket-name, --aws-access-key (or AWS_ACCESS_KEY env. variable), --aws-secret-key (or AWS_SECRET_Key env.variable)) cannot be " +
+                "AWS S3 connection (--aws-bucket-name, --aws-access-key (or AWS_ACCESS_KEY_ID env. variable), --aws-secret-key (or AWS_SECRET_ACCESS_KEY env.variable)) cannot be " +
                 "specified together.");
         }
 
@@ -488,17 +525,47 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         {
             if (!GetAwsAccessKey(args).HasValue())
             {
-                throw new OctoshiftCliException("Either --aws-access-key or AWS_ACCESS_KEY environment variable must be set.");
+#pragma warning disable CS0618
+                if (_environmentVariableProvider.AwsAccessKey(false).HasValue())
+#pragma warning restore CS0618
+                {
+                    _log.LogWarning("AWS_ACCESS_KEY environment variable is deprecated and will be removed in future releases. Please consider using AWS_ACCESS_KEY_ID environment variable instead.");
+                }
+                else
+                {
+                    throw new OctoshiftCliException("Either --aws-access-key or AWS_ACCESS_KEY_ID environment variable must be set.");
+                }
             }
 
             if (!GetAwsSecretKey(args).HasValue())
             {
-                throw new OctoshiftCliException("Either --aws-secret-key or AWS_SECRET_KEY environment variable must be set.");
+#pragma warning disable CS0618
+                if (_environmentVariableProvider.AwsSecretKey(false).HasValue())
+#pragma warning restore CS0618
+                {
+                    _log.LogWarning("AWS_SECRET_KEY environment variable is deprecated and will be removed in future releases. Please consider using AWS_SECRET_ACCESS_KEY environment variable instead.");
+                }
+                else
+                {
+                    throw new OctoshiftCliException("Either --aws-secret-key or AWS_SECRET_ACCESS_KEY environment variable must be set.");
+                }
+            }
+
+            if (GetAwsSessionToken(args).HasValue() && GetAwsRegion(args).IsNullOrWhiteSpace())
+            {
+                throw new OctoshiftCliException(
+                    "--aws-region or AWS_REGION environment variable must be provided with --aws-session-token or AWS_SESSION_TOKEN environment variable.");
+            }
+
+            if (!GetAwsRegion(args).HasValue())
+            {
+                _log.LogWarning("Specifying an AWS region with the --aws-region argument or AWS_REGION environment variable is currently not required, " +
+                                "but will be required in a future release. Defaulting to us-east-1.");
             }
         }
-        else if (args.AwsAccessKey.HasValue() || args.AwsSecretKey.HasValue())
+        else if (new[] { args.AwsAccessKey, args.AwsSecretKey, args.AwsSessionToken, args.AwsRegion }.Any(x => x.HasValue()))
         {
-            throw new OctoshiftCliException("--aws-access-key and --aws-secret-key can only be provided with --aws-bucket-name.");
+            throw new OctoshiftCliException("The AWS S3 bucket name must be provided with --aws-bucket-name if other AWS S3 upload options are set.");
         }
     }
 
@@ -515,9 +582,14 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         }
     }
 
-    private string GetAwsAccessKey(MigrateRepoCommandArgs args) => args.AwsAccessKey.HasValue() ? args.AwsAccessKey : _environmentVariableProvider.AwsAccessKey(false);
+    private string GetAwsAccessKey(MigrateRepoCommandArgs args) => args.AwsAccessKey.HasValue() ? args.AwsAccessKey : _environmentVariableProvider.AwsAccessKeyId(false);
 
-    private string GetAwsSecretKey(MigrateRepoCommandArgs args) => args.AwsSecretKey.HasValue() ? args.AwsSecretKey : _environmentVariableProvider.AwsSecretKey(false);
+    private string GetAwsSecretKey(MigrateRepoCommandArgs args) => args.AwsSecretKey.HasValue() ? args.AwsSecretKey : _environmentVariableProvider.AwsSecretAccessKey(false);
+
+    private string GetAwsRegion(MigrateRepoCommandArgs args) => args.AwsRegion.HasValue() ? args.AwsRegion : _environmentVariableProvider.AwsRegion(false);
+
+    private string GetAwsSessionToken(MigrateRepoCommandArgs args) =>
+        args.AwsSessionToken.HasValue() ? args.AwsSessionToken : _environmentVariableProvider.AwsSessionToken(false);
 
     private string GetAzureStorageConnectionString(MigrateRepoCommandArgs args) => args.AzureStorageConnectionString.HasValue()
         ? args.AzureStorageConnectionString