+++ src/gei/Handlers/MigrateRepoCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
@@ -17,13 +18,14 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
     private readonly AwsApi _awsApi;
     private readonly EnvironmentVariableProvider _environmentVariableProvider;
     private readonly HttpDownloadService _httpDownloadService;
+    private readonly FileSystemProvider _fileSystemProvider;
     private const int ARCHIVE_GENERATION_TIMEOUT_IN_HOURS = 10;
     private const int CHECK_STATUS_DELAY_IN_MILLISECONDS = 10000; // 10 seconds
     private const string GIT_ARCHIVE_FILE_NAME = "git_archive.tar.gz";
     private const string METADATA_ARCHIVE_FILE_NAME = "metadata_archive.tar.gz";
     private const string DEFAULT_GITHUB_BASE_URL = "https://github.com";
 
-    public MigrateRepoCommandHandler(OctoLogger log, GithubApi sourceGithubApi, GithubApi targetGithubApi, EnvironmentVariableProvider environmentVariableProvider, AzureApi azureApi, AwsApi awsApi, HttpDownloadService httpDownloadService)
+    public MigrateRepoCommandHandler(OctoLogger log, GithubApi sourceGithubApi, GithubApi targetGithubApi, EnvironmentVariableProvider environmentVariableProvider, AzureApi azureApi, AwsApi awsApi, HttpDownloadService httpDownloadService, FileSystemProvider fileSystemProvider)
     {
         _log = log;
         _sourceGithubApi = sourceGithubApi;
@@ -32,6 +34,7 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         _azureApi = azureApi;
         _awsApi = awsApi;
         _httpDownloadService = httpDownloadService;
+        _fileSystemProvider = fileSystemProvider;
     }
 
     public async Task Handle(MigrateRepoCommandArgs args)
@@ -46,6 +49,9 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         _log.RegisterSecret(args.GithubSourcePat);
         _log.RegisterSecret(args.GithubTargetPat);
         _log.RegisterSecret(args.AzureStorageConnectionString);
+        _log.RegisterSecret(args.AwsAccessKey);
+        _log.RegisterSecret(args.AwsSecretKey);
+        _log.RegisterSecret(args.AwsSessionToken);
 
         LogOptions(args);
 
@@ -124,22 +130,24 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
             return;
         }
 
-        var (migrationState, _, failureReason) = await _targetGithubApi.GetMigration(migrationId);
+        var (migrationState, _, failureReason, migrationLogUrl) = await _targetGithubApi.GetMigration(migrationId);
 
         while (RepositoryMigrationStatus.IsPending(migrationState))
         {
             _log.LogInformation($"Migration in progress (ID: {migrationId}). State: {migrationState}. Waiting 10 seconds...");
             await Task.Delay(10000);
-            (migrationState, _, failureReason) = await _targetGithubApi.GetMigration(migrationId);
+            (migrationState, _, failureReason, migrationLogUrl) = await _targetGithubApi.GetMigration(migrationId);
         }
 
         if (RepositoryMigrationStatus.IsFailed(migrationState))
         {
             _log.LogError($"Migration Failed. Migration ID: {migrationId}");
+            _log.LogInformation($"Migration log available at {migrationLogUrl} or by running `gh {CliContext.RootCommand} download-logs --github-target-org {args.GithubTargetOrg} --target-repo {args.TargetRepo}`");
             throw new OctoshiftCliException(failureReason);
         }
 
         _log.LogSuccess($"Migration completed (ID: {migrationId})! State: {migrationState}");
+        _log.LogInformation($"Migration log available at {migrationLogUrl} or by running `gh {CliContext.RootCommand} download-logs --github-target-org {args.GithubTargetOrg} --target-repo {args.TargetRepo}`");
     }
 
     private string GetSourceToken(MigrateRepoCommandArgs args) =>
@@ -198,18 +206,49 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
             return (gitArchiveUrl, metadataArchiveUrl);
         }
 
-        _log.LogInformation($"Downloading archive from {gitArchiveUrl}");
-        var gitArchiveContent = await _httpDownloadService.DownloadToBytes(gitArchiveUrl);
+        var gitArchiveFilePath = _fileSystemProvider.GetTempFileName();
+        var metadataArchiveFilePath = _fileSystemProvider.GetTempFileName();
+        try
+        {
+            _log.LogInformation($"Downloading archive from {gitArchiveUrl}");
+            await _httpDownloadService.DownloadToFile(gitArchiveUrl, gitArchiveFilePath);
 
-        _log.LogInformation($"Downloading archive from {metadataArchiveUrl}");
-        var metadataArchiveContent = await _httpDownloadService.DownloadToBytes(metadataArchiveUrl);
+            _log.LogInformation($"Downloading archive from {metadataArchiveUrl}");
+            await _httpDownloadService.DownloadToFile(metadataArchiveUrl, metadataArchiveFilePath);
 
-        return _awsApi.HasValue() ?
-            await UploadArchivesToAws(awsBucketName, gitArchiveFileName, gitArchiveContent, metadataArchiveFileName, metadataArchiveContent) :
-            await UploadArchivesToAzure(gitArchiveFileName, gitArchiveContent, metadataArchiveFileName, metadataArchiveContent);
+#pragma warning disable IDE0063
+            await using (var gitArchiveContent = _fileSystemProvider.OpenRead(gitArchiveFilePath))
+            await using (var metadataArchiveContent = _fileSystemProvider.OpenRead(metadataArchiveFilePath))
+#pragma warning restore IDE0063
+            {
+                return _awsApi.HasValue()
+                    ? await UploadArchivesToAws(awsBucketName, gitArchiveFileName, gitArchiveContent, metadataArchiveFileName, metadataArchiveContent)
+                    : await UploadArchivesToAzure(gitArchiveFileName, gitArchiveContent, metadataArchiveFileName, metadataArchiveContent);
+            }
+        }
+        finally
+        {
+            DeleteArchive(gitArchiveFilePath);
+            DeleteArchive(metadataArchiveFilePath);
+        }
     }
 
-    private async Task<(string, string)> UploadArchivesToAzure(string gitArchiveFileName, byte[] gitArchiveContent, string metadataArchiveFileName, byte[] metadataArchiveContent)
+    private void DeleteArchive(string path)
+    {
+        try
+        {
+            _fileSystemProvider.DeleteIfExists(path);
+        }
+#pragma warning disable CA1031
+        catch (Exception ex)
+#pragma warning restore CA1031
+        {
+            _log.LogWarning($"Couldn't delete the downloaded archive at \"{path}\". Error message: \"{ex.Message}\"");
+            _log.LogVerbose(ex.ToString());
+        }
+    }
+
+    private async Task<(string, string)> UploadArchivesToAzure(string gitArchiveFileName, Stream gitArchiveContent, string metadataArchiveFileName, Stream metadataArchiveContent)
     {
         _log.LogInformation($"Uploading archive {gitArchiveFileName} to Azure Blob Storage");
         var authenticatedGitArchiveUri = await _azureApi.UploadToBlob(gitArchiveFileName, gitArchiveContent);
@@ -219,7 +258,7 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         return (authenticatedGitArchiveUri.ToString(), authenticatedMetadataArchiveUri.ToString());
     }
 
-    private async Task<(string, string)> UploadArchivesToAws(string bucketName, string gitArchiveFileName, byte[] gitArchiveContent, string metadataArchiveFileName, byte[] metadataArchiveContent)
+    private async Task<(string, string)> UploadArchivesToAws(string bucketName, string gitArchiveFileName, Stream gitArchiveContent, string metadataArchiveFileName, Stream metadataArchiveContent)
     {
         _log.LogInformation($"Uploading archive {gitArchiveFileName} to AWS S3");
         var authenticatedGitArchiveUri = await _awsApi.UploadToBucket(bucketName, gitArchiveContent, gitArchiveFileName);
@@ -240,10 +279,6 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
             {
                 return await githubApi.GetArchiveMigrationUrl(githubSourceOrg, archiveId);
             }
-            if (archiveStatus == ArchiveMigrationStatus.Failed)
-            {
-                throw new OctoshiftCliException($"Archive generation failed for id: {archiveId}");
-            }
             await Task.Delay(CHECK_STATUS_DELAY_IN_MILLISECONDS);
         }
         throw new TimeoutException($"Archive generation timed out after {ARCHIVE_GENERATION_TIMEOUT_IN_HOURS} hours");
@@ -251,18 +286,26 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
 
     private async Task<bool> DetermineIfBlobCredentialsRequired(MigrateRepoCommandArgs args)
     {
-        var blobCredentialsRequired = true;
+        var blobCredentialsRequired = false;
+
         if (args.GhesApiUrl.HasValue())
         {
+            blobCredentialsRequired = true;
+
             _log.LogInformation("Using GitHub Enterprise Server - verifying server version");
             var ghesVersion = await _sourceGithubApi.GetEnterpriseServerVersion();
 
             if (ghesVersion != null)
             {
                 _log.LogInformation($"GitHub Enterprise Server version {ghesVersion} detected");
-                if (new Version(ghesVersion) >= new Version(3, 8, 0))
+
+                if (Version.TryParse(ghesVersion, out var parsedVersion))
                 {
-                    blobCredentialsRequired = false;
+                    blobCredentialsRequired = parsedVersion < new Version(3, 8, 0);
+                }
+                else
+                {
+                    _log.LogInformation($"Unable to parse the version number, defaulting to using CLI for blob storage uploads");
                 }
             }
         }
@@ -368,6 +411,11 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
             _log.LogInformation("LOCK SOURCE REPO: true");
         }
 
+        LogAwsOptions(args);
+    }
+
+    private void LogAwsOptions(MigrateRepoCommandArgs args)
+    {
         if (args.AwsBucketName.HasValue())
         {
             _log.LogInformation($"AWS BUCKET NAME: {args.AwsBucketName}");
@@ -382,7 +430,18 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         {
             _log.LogInformation("AWS SECRET KEY: ***");
         }
+
+        if (args.AwsSessionToken.HasValue())
+        {
+            _log.LogInformation("AWS SESSION TOKEN: ***");
+        }
+
+        if (args.AwsRegion.HasValue())
+        {
+            _log.LogInformation($"AWS REGION: {args.AwsRegion}");
+        }
     }
+
     private void ValidateOptions(MigrateRepoCommandArgs args, bool cloudCredentialsRequired)
     {
         if (args.GithubTargetPat.HasValue() && args.GithubSourcePat.IsNullOrWhiteSpace())
@@ -441,7 +500,7 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
             {
                 throw new OctoshiftCliException(
                     "Either Azure storage connection (--azure-storage-connection-string or AZURE_STORAGE_CONNECTION_STRING env. variable) or " +
-                    "AWS S3 connection (--aws-bucket-name, --aws-access-key (or AWS_ACCESS_KEY env. variable), --aws-secret-key (or AWS_SECRET_KEY env.variable)) " +
+                    "AWS S3 connection (--aws-bucket-name, --aws-access-key (or AWS_ACCESS_KEY_ID env. variable), --aws-secret-key (or AWS_SECRET_ACCESS_KEY env.variable)) " +
                     "must be provided.");
             }
 
@@ -449,7 +508,7 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
             {
                 throw new OctoshiftCliException(
                     "Azure storage connection (--azure-storage-connection-string or AZURE_STORAGE_CONNECTION_STRING env. variable) and " +
-                    "AWS S3 connection (--aws-bucket-name, --aws-access-key (or AWS_ACCESS_KEY env. variable), --aws-secret-key (or AWS_SECRET_Key env.variable)) cannot be " +
+                    "AWS S3 connection (--aws-bucket-name, --aws-access-key (or AWS_ACCESS_KEY_ID env. variable), --aws-secret-key (or AWS_SECRET_ACCESS_KEY env.variable)) cannot be " +
                     "specified together.");
             }
 
@@ -457,17 +516,47 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
             {
                 if (!GetAwsAccessKey(args).HasValue())
                 {
-                    throw new OctoshiftCliException("Either --aws-access-key or AWS_ACCESS_KEY environment variable must be set.");
+#pragma warning disable CS0618
+                    if (_environmentVariableProvider.AwsAccessKey(false).HasValue())
+#pragma warning restore CS0618
+                    {
+                        _log.LogWarning("AWS_ACCESS_KEY environment variable is deprecated and will be removed in future releases. Please consider using AWS_ACCESS_KEY_ID environment variable instead.");
+                    }
+                    else
+                    {
+                        throw new OctoshiftCliException("Either --aws-access-key or AWS_ACCESS_KEY_ID environment variable must be set.");
+                    }
                 }
 
                 if (!GetAwsSecretKey(args).HasValue())
                 {
-                    throw new OctoshiftCliException("Either --aws-secret-key or AWS_SECRET_KEY environment variable must be set.");
+#pragma warning disable CS0618
+                    if (_environmentVariableProvider.AwsSecretKey(false).HasValue())
+#pragma warning restore CS0618
+                    {
+                        _log.LogWarning("AWS_SECRET_KEY environment variable is deprecated and will be removed in future releases. Please consider using AWS_SECRET_ACCESS_KEY environment variable instead.");
+                    }
+                    else
+                    {
+                        throw new OctoshiftCliException("Either --aws-secret-key or AWS_SECRET_ACCESS_KEY environment variable must be set.");
+                    }
+                }
+
+                if (GetAwsSessionToken(args).HasValue() && GetAwsRegion(args).IsNullOrWhiteSpace())
+                {
+                    throw new OctoshiftCliException(
+                        "--aws-region or AWS_REGION environment variable must be provided with --aws-session-token or AWS_SESSION_TOKEN environment variable.");
+                }
+
+                if (!GetAwsRegion(args).HasValue())
+                {
+                    _log.LogWarning("Specifying an AWS region with the --aws-region argument or AWS_REGION environment variable is currently not required, " +
+                                    "but will be required in a future release. Defaulting to us-east-1.");
                 }
             }
-            else if (args.AwsAccessKey.HasValue() || args.AwsSecretKey.HasValue())
+            else if (new[] { args.AwsAccessKey, args.AwsSecretKey, args.AwsSessionToken, args.AwsRegion }.Any(x => x.HasValue()))
             {
-                throw new OctoshiftCliException("--aws-access-key and --aws-secret-key can only be provided with --aws-bucket-name.");
+                throw new OctoshiftCliException("The AWS S3 bucket name must be provided with --aws-bucket-name if other AWS S3 upload options are set.");
             }
         }
         else
@@ -484,9 +573,14 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         }
     }
 
-    private string GetAwsAccessKey(MigrateRepoCommandArgs args) => args.AwsAccessKey.HasValue() ? args.AwsAccessKey : _environmentVariableProvider.AwsAccessKey(false);
+    private string GetAwsAccessKey(MigrateRepoCommandArgs args) => args.AwsAccessKey.HasValue() ? args.AwsAccessKey : _environmentVariableProvider.AwsAccessKeyId(false);
 
-    private string GetAwsSecretKey(MigrateRepoCommandArgs args) => args.AwsSecretKey.HasValue() ? args.AwsSecretKey : _environmentVariableProvider.AwsSecretKey(false);
+    private string GetAwsSecretKey(MigrateRepoCommandArgs args) => args.AwsSecretKey.HasValue() ? args.AwsSecretKey : _environmentVariableProvider.AwsSecretAccessKey(false);
+
+    private string GetAwsRegion(MigrateRepoCommandArgs args) => args.AwsRegion.HasValue() ? args.AwsRegion : _environmentVariableProvider.AwsRegion(false);
+
+    private string GetAwsSessionToken(MigrateRepoCommandArgs args) =>
+        args.AwsSessionToken.HasValue() ? args.AwsSessionToken : _environmentVariableProvider.AwsSessionToken(false);
 
     private string GetAzureStorageConnectionString(MigrateRepoCommandArgs args) => args.AzureStorageConnectionString.HasValue()
         ? args.AzureStorageConnectionString